// src/enums/OrderTypes.ts
export enum OrderTypes{
  EditClient ="Edycja danych klienta",
  CreateUser = "Utworzenie użytkownika",
  EditUser = "Modyfikacja użytkownika",
  EditEmployee = "Modyfikacja danych pracownika",
  BlockCreditCard = "Zablokowanie karty kredytowej",
  UnblockCreditCard = "Odblokowanie karty kredytowej",
  CreateCreditCard = "Wyrób nowej karty kredytowej",
  DeleteCreditCard = "Wycofanie karty kredytowej",
  LoanRequest = "Podanie o kredyt"
}